Simplified the GAN implementation by leaving only gan_tf, replacing gan_components with DenseNets, and, just like with the configs, distributed the corresponding parameter_spaces into their folders.
