+++ degann/networks/topology/GAN/tf_gan.py
@@ -1,23 +1,20 @@
-from typing import Optional, Dict, Callable, List
+from typing import Optional, Dict, Callable, List, Any, Self, Tuple
 import tensorflow as tf
 from tensorflow import keras
 
-from degann.networks.topology.configs import GANConfig
-from degann.networks.topology.tf_gan_components import (
-    TensorflowGenerator,
-    TensorflowDiscriminator,
-)
-from degann.networks import losses, metrics
+from degann.networks.topology.DenseNet.tf_densenet import TensorflowDenseNet
+from degann.networks.topology.GAN.config import GANConfig
+from degann.networks import losses, metrics, optimizers
 
 
 class TensorflowGAN(tf.keras.Model):
     def __init__(
         self,
         config: GANConfig,
-        gen_weight_init=keras.initializers.HeNormal(),
-        gen_bias_init=tf.zeros_initializer(),
-        disc_weight_init=keras.initializers.HeNormal(),
-        disc_bias_init=tf.zeros_initializer(),
+        gen_weight=keras.initializers.HeNormal(),
+        gen_biases=keras.initializers.Zeros(),
+        disc_weight=keras.initializers.HeNormal(),
+        disc_biases=keras.initializers.Zeros(),
         is_debug: bool = False,
         **kwargs,
     ):
@@ -26,25 +23,18 @@ class TensorflowGAN(tf.keras.Model):
         self.config = config
         self.is_debug = is_debug
 
-        self.generator = TensorflowGenerator(
-            input_size=config.gen_input_size,
-            output_size=config.gen_output_size,
-            block_sizes=config.gen_block_sizes,
-            activation_funcs=config.gen_activation_funcs,
-            out_activation=config.gen_out_activation,
-            weight_init=gen_weight_init,
-            bias_init=gen_bias_init,
+        self.generator = TensorflowDenseNet(
+            config=config.gen_config,
+            weight=gen_weight,
+            biases=gen_biases,
             is_debug=is_debug,
             name="TFGenerator",
         )
 
-        self.discriminator = TensorflowDiscriminator(
-            input_size=config.gen_input_size + config.gen_output_size,
-            output_size=1,
-            block_sizes=config.disc_block_sizes,
-            activation_funcs=config.disc_activation_funcs,
-            weight_init=disc_weight_init,
-            bias_init=disc_bias_init,
+        self.discriminator = TensorflowDenseNet(
+            config=config.disc_config,
+            weight=disc_weight,
+            biases=disc_biases,
             is_debug=is_debug,
             name="TFDiscriminator",
         )
@@ -57,25 +47,76 @@ class TensorflowGAN(tf.keras.Model):
         self.metric_fake = tf.keras.metrics.Mean(name="metric_disc_fake")
         self.trained_time = {"train_time": 0.0, "epoch_time": [], "predict_time": 0}
 
-    def compile(
-        self,
-        gen_optimizer: str = "Adam",
-        disc_optimizer: str = "Adam",
-        gen_loss_func: str = "MeanSquaredError",
-        disc_loss_func: str = "MeanSquaredError",
-        **kwargs,
+    def custom_compile(
+            self,
+            gen_rate=1e-2,
+            disc_rate=1e-2,
+            gen_optimizer="Adam",
+            disc_optimizer="Adam",
+            gen_loss_func="MeanSquaredError",
+            disc_loss_func="MeanSquaredError",
+            metric_funcs=None,
+            run_eagerly=False,
     ):
-        super().compile(optimizer="sgd", loss="mse", run_eagerly=True, **kwargs)
+        """
+        Configures the model for training
 
-        self.gen_optimizer = keras.optimizers.get(gen_optimizer)
-        self.disc_optimizer = keras.optimizers.get(disc_optimizer)
+        Parameters
+        ----------
+        gen_rate: float
+            learning rate for generator optimizer
+        disc_rate: float
+            learning rate for discriminator optimizer
+        gen_optimizer: str
+            name of generator optimizer
+        disc_optimizer: str
+            name of discriminator optimizer
+        gen_loss_func: str
+            name of generator loss function
+        disc_loss_func: str
+            name of discriminator loss function
+        metric_funcs: list[str]
+            list with metric function names
+        run_eagerly: bool
+
+        Returns
+        -------
+
+        """
+
+        if metric_funcs is None:
+            metric_funcs = []
+
+        self.gen_optimizer = optimizers.get_optimizer(gen_optimizer)(learning_rate=gen_rate)
+        self.disc_optimizer = optimizers.get_optimizer(disc_optimizer)(learning_rate=disc_rate)
         self.gen_loss = losses.get_loss(gen_loss_func)
         self.disc_loss = losses.get_loss(disc_loss_func)
+        m = [metrics.get_metric(metric) for metric in metric_funcs]
+        self.compile(
+            optimizer="sgd",
+            loss="mse",
+            metrics=m,
+            run_eagerly=run_eagerly,
+        )
 
-    def call(self, inputs, **kwargs):
+    def call(self, inputs, **kwargs) -> tf.Tensor:
+        """Forward pass through generator."""
         return self.generator(inputs, **kwargs)
 
-    def train_step(self, data):
+    def train_step(self, data) -> Dict[str, tf.Tensor]:
+        """
+        Custom train step for GAN.
+
+        Parameters
+        ----------
+        data : tuple
+            Pair of (x, y_true).
+
+        Returns
+        -------
+        dict
+            Dictionary with loss values and metrics.
+        """
         x, y_true = data
 
         with tf.GradientTape() as disc_tape:
@@ -127,9 +168,19 @@ class TensorflowGAN(tf.keras.Model):
         }
 
     def set_name(self, name):
-        self.name = name
+        """Set model name."""
+        self._name = name
 
-    def to_dict(self, **kwargs):
+    def to_dict(self) -> Dict[str, Any]:
+        """
+        Export neural network to dictionary.
+
+        Returns
+        -------
+        dict
+            Dictionary representation of the model.
+        """
+
         res = {
             "net_type": "TFGAN",
             "name": self.name,
@@ -140,7 +191,20 @@ class TensorflowGAN(tf.keras.Model):
         return res
 
     @classmethod
-    def from_dict(cls, config_dict: Dict, **kwargs):
+    def from_dict(cls, config_dict: Dict[str, Any], **kwargs) -> Self:
+        """
+        Restore neural network from dictionary.
+
+        Parameters
+        ----------
+        config_dict : dict
+            Dictionary with model configuration.
+
+        Returns
+        -------
+        TensorflowGAN
+            Restored model instance.
+        """
         config = GANConfig.from_dict(config_dict["config"])
         model = cls(config=config, **kwargs)
         model.generator.from_dict(config_dict["generator"])
@@ -148,7 +212,15 @@ class TensorflowGAN(tf.keras.Model):
         return model
 
     @property
-    def get_activations(self) -> tuple[list, list]:
+    def get_activations(self) -> Tuple[List[str], List[str]]:
+        """
+        Get list of activation functions for each layer.
+
+        Returns
+        -------
+        tuple[list, list]
+            Activations for generator and discriminator.
+        """
         gen_activations = self.generator.get_activations
         disc_activations = self.discriminator.get_activations
         return gen_activations, disc_activations
