+++ degann/networks/topology/__init__.py
@@ -1,1 +1,1 @@
-from degann.networks.topology.tf_densenet import *
+from degann.networks.topology.DenseNet.tf_densenet import *
