+++ tests/test_simple_equation.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
 from degann.networks.imodel import IModel
-from degann.networks.topology.configs import DenseNetConfig
+from degann.networks.topology.DenseNet.config import DenseNetConfig
 from tests.utils import array_compare
 from degann.equations import simple_equation
